Move `update-password` route from `(protected)` to `(auth)` and update associated logic and navigating back to login page after success.
